Investigate basic graph structure
The investigations revealed a graph with only one connected component.
Additionally, the graph seems to be fully connected and also have two
extra edges out of the total possible edges, meaning that it must have
some auto-connections. In that case, maybe there are more, and the graph
is not fully connected at all.

One possible reason for that is the after process of erasing missing
nodes that were present in the control/stages dataframes.

Further investigation on the patients network generation must be perfor-
med. Maybe tinkering with the correlation and p-values thresholds.