Add patients correlation investigation
It turned out that every possible pair of patients had a correlation
above the established thresholds (r>0.8; p<0.05), except for two pairs.
A correlation histogram was plotted and revealed a peak around 0.95~0.96
regarding the r value.

More selective thresholds should be used connecting only the ones that
have a similar profile.

Additionally, for every node, an autoconnection must be made.