Add notebook for filter genes by IQR
Although 0.6 was threshold value chosen for dimensionality reduction in
the other project initially, here the focus is to reduce more the some
thousands of genes. Therefore, I tested some values and found that 3 is
a nice arbitrary threshold that maintained less than 1k genes in the da-
taframe. But this parameter should be revisited and a justification must
be provided.