Test GAT performance using reduced genes dataset version
This reduced version using IQR threshold of 3, resulting in 650 genes,
and the 650 genes themselves as features, seemed to improve performance
somewhat. What really matters: the model is learning and it's more resi-
lient against overfitting (compared to a simple MLP).

Some investigations for the IQR parameter must be done, and also using
all available patients even if they do not connect with anyone besides
themselves.