Add data processing for other features (patient subset + iqr FS)
After getting only the insterested patients from the gene expression
dataset, that is, the ones with valid cancer stage information (597),
the intersection between them and the other datasets, methylation and
miRNA expression data, were performed resulting in 201 patients.

The content of the methylation data did not suffer any transformation
given its ranges lying already between 0 and 1. The miRNA data, in the
other hand, had to be log2(expr+1) normalized.

Furthermore, an IQR feature selection was performed were only the fea-
tures with the top k biggest IQR were selected. For gene expression data
the k was set to 500, and for the methylation and miRNA data the k was
set to 100.